HUM-89 - folder rename utils to helpers
