+++ packages/apps/human-app/frontend/src/modules/worker/components/jobs-discovery/helpers/get-oracles-table-columns.tsx
