+++ packages/apps/human-app/frontend/src/modules/worker/components/jobs-discovery/helpers/handle-hcaptcha-navigation.ts
