+++ packages/apps/human-app/frontend/src/modules/worker/components/jobs-discovery/helpers/is-hcaptcha-oracle.tsx
