+++ packages/apps/human-app/frontend/src/modules/worker/components/jobs-discovery/helpers/navigate-to-registration.ts
