+++ packages/apps/human-app/frontend/src/modules/worker/components/jobs-discovery/helpers/should-navigate-to-registration.ts
